revert resize event parameters
There's not much that can be done about the initial size of the chart, the resize is doing it's job properly (if there's any space for the container to expand).
Since the initial height of the container is 0px, the best approach is to ensure the element in which the chart is nested is growing to the required size, even without the chart in it.